refactor(/src,index.html):MVC패턴 적용
model,view,controller폴더 생성 후 각 역할에 맞는 함수 위치 변경,할당
view,model의 경우 클라스 문법을 사용하여 코드의 모듈화에 집중함
전체적인 코드의 구조변화 리펙터링임에 주의하여 코드구조변경 외의 다른 기능추가나
다른 영역 리펙터링을 하지 않음
