+++ README.md
@@ -25,6 +25,11 @@
   8.1 변수재설정
   8.2 ui초기화
 
+## MVC 패턴 적용 로드맵
+1.기존의 1주차 코드에서 model,controller,view폴더로 나누어 
+역활에 따라 class내 함수 재배치
+2.eslint에 따른 코드컨벤션 적용
+3.추가 개선점 확인 밎 리팩터링 진행
 
 ## 🎯 기능 요구사항
