+++ src/utils/additionalUtilFunction.js
