+++ src/utils/checkInput.js
